reports stored as individual facts
Rather than being stored with serialized facts in a serialized report,
individual facts are stored and replaced with their GUID's.  Reports
are stored with GUIDs and contents are reconstituted into facts upon
extraction.

This is ugly code that depends on the particular form of data structure
returned by $fact->as_struct().